rtf_core::graphos::platform_query

Trait Client

pub trait Client {
    // Required method
    async fn post_operation(
        &self,
        body: &impl Serialize,
    ) -> Result<Value, Error>;

    // Provided method
    async fn execute_operation<T>(
        &self,
        variables: T::Variables,
    ) -> Result<T::ResponseData, Error>
       where T: GraphQLQuery { ... }
}

An API client that can make requests to the Apollo platform API.

Required Methods

async fn post_operation(&self, body: &impl Serialize) -> Result<Value, Error>

POST a GraphQL operation to the studio API with appropriate headers, returning the raw JSON response.

Client::execute_operation is used to handle creating the POST body and parsing the response.

Provided Methods

async fn execute_operation<T>( &self, variables: T::Variables, ) -> Result<T::ResponseData, Error>
where T: GraphQLQuery,

Helper for making an API request to studio and handling any serialization or graphQL errors so that implementations of PlatformQuery::try_parse only need to care about mapping valid data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.