rtf_core::graphos::platform_query

Trait PlatformQuery

pub trait PlatformQuery: GraphQLQuery + Sized
where
    Self::Variables: Clone,
{
    type Output;
    type Error: From<Error>;

    // Required method
    fn try_parse(
        data: Self::ResponseData,
        variables: Self::Variables,
    ) -> Result<Self::Output, Self::Error>;

    // Provided method
    async fn fetch(
        variables: Self::Variables,
        client: &impl Client,
    ) -> Result<Self::Output, Self::Error> { ... }
}

Helper trait for packaging up making a graphql request to the Apollo platform API and parsing the response into a more ergonomic type.

Deriving GraphQLQuery

Types that implement this trait need to derive the [GraphQLQuery] trait from [graphql_client] first. The schema file for the platform API (engine@prod) along with the query files used to define new queries are located in the top level resources directory of this crate. The details on how to work with the graphql_client crate are covered in their docs but there are couple of things we need to specify as boiler plate in order for the PlatformQuery trait to work:

variables need to derive Clone.

responses need to derive Deserialize.

The name of the query in your graphql file needs to exactly match the name of the Rust struct

The macro will then generate a module with a name that is the snake_case transform of your struct (so MyQuery becomes my_query) that contains all of the generated types needed to make your request and parse the response.

Example

# contents of queries/my-query.graphql
query MyQuery(
  $my_arg: String!,
  $my_other_arg: Int,
  $my_enums: [MyEnum]!
) {
  foo(arg: $my_arg) {
    bar(arg: $my_other_arg, enums: $my_enums) {
      myData # a nullable string
     }
  }
}

#[derive(GraphQLQuery)]
#[graphql(
    schema_path = "resources/engine-prod-schema.graphql",
    query_path = "queries/my-query.graphql",
    response_derives = "Deserialize",
    variables_derives = "Clone"
)]
pub struct MyQuery;

impl PlatformQuery for MyQuery {
    type Output = String;
    type Error = &'static str;

    fn try_parse(
        data: Self::ResponseData,
        _vars: my_query::Variables
    ) -> Result<String, &'static str> {
        // The nested structure of Self::ResponseData matches the your graphQL query
        let maybe_my_data = data.foo.bar.my_data;

        match maybe_my_data {
            Some(s) => Ok(s),
            None => Err("myData was null"),
        }
    }
}

// Making a request
let vars = my_query::Variables {
    my_arg: "some arg".to_string(),      // non-nullable types in graphql map to Rust types
    my_other_arg: Some(42),              // nullable types map to options
    my_enums: vec![my_query::MyEnum::A], // enums are available in the generated module
};
let api_key = "...";
let client = ReqwestClient::new_prod(api_key);

let parsed_response = MyQuery::fetch(vars, &client).await?;

Required Associated Types

type Output

The output type returned from try_parse

type Error: From<Error>

The error type returned from try_parse

Required Methods

fn try_parse( data: Self::ResponseData, variables: Self::Variables, ) -> Result<Self::Output, Self::Error>

Attempt to parse the raw data returned from a graphQL query into a usable Rust type.

Provided Methods

async fn fetch( variables: Self::Variables, client: &impl Client, ) -> Result<Self::Output, Self::Error>

Execute this query and parse the returned data

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.