rtf_core::graphos::supergraph::operations

Module canned_operations

Pull top operation signatures from studio for a given supergraph and rewrite them in order to generate canned graphQL requests that are valid for the parent supergraph schema.

The raw operation details pulled from Studio are redacted in a number of ways that cause them to become invalid. See Signature::parse_and_fix for details of what issues we currently support fixing.

Structs

CannedOperation

A canned operation is a validated graphQL operation from Studio paired with generated data to use as the variables required to send the request to a running Router.

Signature

The redacted signature for a given graphQL operation stored in Studio.

Functions

canned_ops_for_ids

For a given supergraph, pull the requested operations by ID and generate canned operation data to be able to use them in requests to a running Router.

top_studio_canned_ops

For a given supergraph, pull the top n operations as reported by studio and generate canned operation data to be able to use them in requests to a running Router.